Split AuthenticationService in RestService and AuthenticationService extend it
